Mostly finished all logic for i5, lastly need to check if deep copy of env actually works.

+++ src/NodeFactNum.java
@@ -1,4 +1,3 @@
-
 // (C) 2013 Jim Buffenbarger
 // All rights reserved.
 
@@ -7,11 +6,11 @@ public class NodeFactNum extends NodeFact {
     private String num;
 
     public NodeFactNum(String num) {
-	this.num=num;
+        this.num=num;
     }
 
     public double eval(Environment env) throws EvalException {
-	return Integer.parseInt(num);
+        return Double.parseDouble(num);
     }
 
 }
